+++ Modules/Intent.Modules.AspNetCore.Controllers.Interop.MediatR/release-notes.md
@@ -1,6 +1,7 @@
 ﻿﻿### Version 4.0.1
 
 - Update: Catered for scenario for our Auto CRUD implementation to auto-wireup a Query field.
+- Controllers will now return a 404 Not Found response for non-collection null results.
 
 ### Version 4.0.0